refactor: use resourceType enum instead of type
