--- src/common/types/ResourceType.ts
@@ -1,1 +1,0 @@
-export type ResourceType = "user" | "room" | "comment" | "note";
